refactor(local-inference): organize live local model validation
- split vibe prompt and Polished gauntlet coverage into separate files
- add shared live local model helper types for reusable prompt validation
- rename live local model environment variables to remove LFM terminology
